+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - more tox testing framework
 - new directory structure with src for python code
 - new plugin mechanism added
+- plugin execution added